FINERACT-1656 Correlation ID propagation and configuration
