--- fineract-provider/src/test/java/org/apache/fineract/infrastructure/core/filters/MutableHttpServletRequest.java
@@ -1,66 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.fineract.infrastructure.core.filters;
-
-import java.util.Collections;
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
- 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletRequestWrapper;
- 
-final class MutableHttpServletRequest extends HttpServletRequestWrapper {
-    
-    private final Map<String, String> customHeaders;
- 
-    public MutableHttpServletRequest(HttpServletRequest request){
-        super(request);
-        this.customHeaders = new HashMap<String, String>();
-    }
-    
-    public void putHeader(String name, String value){
-        this.customHeaders.put(name, value);
-    }
-
-    @Override
-    public String getHeader(String name) {        
-        String headerValue = customHeaders.get(name);
-        
-        if (headerValue != null){
-            return headerValue;
-        }        
-        return ((HttpServletRequest) getRequest()).getHeader(name);
-    }
- 
-    @Override
-    public Enumeration<String> getHeaderNames() {        
-        Set<String> set = new HashSet<String>(customHeaders.keySet());
-                
-        @SuppressWarnings("unchecked")
-        Enumeration<String> e = ((HttpServletRequest) getRequest()).getHeaderNames();
-        while (e.hasMoreElements()) {        
-            String n = e.nextElement();
-            set.add(n);
-        }
-        return Collections.enumeration(set);
-    }
-}
